Organize explorer file structure
