+++ explorer/src/components/common/Overlay.tsx
@@ -4,6 +4,6 @@ type OverlayProps = {
   show: boolean;
 };
 
-export default function Overlay({ show }: OverlayProps) {
+export function Overlay({ show }: OverlayProps) {
   return <div className={`modal-backdrop fade${show ? " show" : ""}`}></div>;
 }
