+++ explorer/src/components/instruction/stake/InitializeDetailsCard.tsx
@@ -8,7 +8,7 @@ import {
 } from "@solana/web3.js";
 import { InstructionCard } from "../InstructionCard";
 import { UnknownDetailsCard } from "../UnknownDetailsCard";
-import Address from "components/common/Address";
+import { Address } from "components/common/Address";
 
 export function InitializeDetailsCard(props: {
   ix: TransactionInstruction;
