+++ explorer/src/components/instruction/system/AllocateDetailsCard.tsx
@@ -7,7 +7,7 @@ import {
 } from "@solana/web3.js";
 import { InstructionCard } from "../InstructionCard";
 import { UnknownDetailsCard } from "../UnknownDetailsCard";
-import Address from "components/common/Address";
+import { Address } from "components/common/Address";
 
 export function AllocateDetailsCard(props: {
   ix: TransactionInstruction;
