+++ explorer/src/components/instruction/system/AssignWithSeedDetailsCard.tsx
@@ -6,9 +6,9 @@ import {
   SystemInstruction,
 } from "@solana/web3.js";
 import { InstructionCard } from "../InstructionCard";
-import Copyable from "components/Copyable";
+import { Copyable } from "components/common/Copyable";
 import { UnknownDetailsCard } from "../UnknownDetailsCard";
-import Address from "components/common/Address";
+import { Address } from "components/common/Address";
 
 export function AssignWithSeedDetailsCard(props: {
   ix: TransactionInstruction;
